Time: 0 ms (100.00%) | Memory: 2.1 MB (60.00%) - LeetSync

+++ 3172-divisible-and-non-divisible-sums-difference/divisible-and-non-divisible-sums-difference.rs
@@ -11,6 +11,6 @@ impl Solution {
             }
         }
 
-        (num1 - num2) as i32
+        num1 - num2
     }
 }
